FUNC: create content factory class.
