+++ src/md_html/elements/__init__.py
@@ -8,7 +8,9 @@ All elements inherit from the BaseElement.
 
 # ruff: noqa: F401
 
+from md_html.elements.base_element import Elements
 from md_html.elements.body import Body
+from md_html.elements.content import Content
 from md_html.elements.head import Head
 from md_html.elements.html import HTML
 from md_html.elements.paragraph import P
