+++ src/md_html/elements/body.py
@@ -1,7 +1,7 @@
 """The Body of the HTML document."""
 
 from md_html.elements.base_element import Elements, HTMLBaseClass
-from md_html.elements.paragraph import P
+from md_html.elements.content import Content
 
 
 class Body(HTMLBaseClass):
@@ -9,10 +9,10 @@ class Body(HTMLBaseClass):
 
     kind = Elements.body
 
-    def __post_init__(self, sections: list[str], **kwargs) -> None:
+    def __post_init__(self, sections: list[Content], *args, **kwargs) -> None:
         """Instantiate the Body class."""
         for section in sections:
             self.content = "\n    ".join(
-                [self.content, P(content=section, **kwargs).html]
+                [self.content, section.build(*args, **kwargs).html]
             )
         self.content = self.content + "\n  "
