--- src/md_html/utils.py
@@ -1,12 +1,0 @@
-"""Utility methods the package."""
-
-from enum import StrEnum
-
-
-class Elements(StrEnum):
-    """HTML elements lookup."""
-
-    paragraph = "p"
-    head = "head"
-    html = "html"
-    body = "body"
